item resource tests
